+++ Content.Server/_CP14/AuraImprint/CP14AuraImprintSystem.cs
@@ -1,5 +1,8 @@
 using Content.Shared._CP14.AuraDNA;
+using Content.Shared._CP14.MagicVision;
+using Content.Shared.Mobs;
 using Robust.Shared.Random;
+using Robust.Shared.Utility;
 
 namespace Content.Server._CP14.AuraImprint;
 
@@ -9,11 +12,13 @@ namespace Content.Server._CP14.AuraImprint;
 public sealed partial class CP14AuraImprintSystem : CP14SharedAuraImprintSystem
 {
     [Dependency] private readonly IRobustRandom _random = default!;
+    [Dependency] private readonly CP14SharedMagicVisionSystem _vision = default!;
     public override void Initialize()
     {
         base.Initialize();
 
         SubscribeLocalEvent<CP14AuraImprintComponent, MapInitEvent>(OnMapInit);
+        SubscribeLocalEvent<CP14AuraImprintComponent, MobStateChangedEvent>(OnMobStateChanged);
     }
 
     private void OnMapInit(Entity<CP14AuraImprintComponent> ent, ref MapInitEvent args)
@@ -33,5 +38,32 @@ public sealed partial class CP14AuraImprintSystem : CP14SharedAuraImprintSystem
         }
 
         return $"[color={ent.Comp.ImprintColor.ToHex()}]{imprint}[/color]";
+    }
+
+    private void OnMobStateChanged(Entity<CP14AuraImprintComponent> ent, ref MobStateChangedEvent args)
+    {
+        switch (args.NewMobState)
+        {
+            case MobState.Critical:
+            {
+                _vision.SpawnMagicVision(
+                    Transform(ent).Coordinates,
+                    new SpriteSpecifier.Rsi(new ResPath("_CP14/Actions/Spells/misc.rsi"), "skull"),
+                    Loc.GetString("cp14-magic-vision-crit"),
+                    TimeSpan.FromMinutes(10),
+                    ent);
+                break;
+            }
+            case MobState.Dead:
+            {
+                _vision.SpawnMagicVision(
+                    Transform(ent).Coordinates,
+                    new SpriteSpecifier.Rsi(new ResPath("_CP14/Actions/Spells/misc.rsi"), "skull_red"),
+                    Loc.GetString("cp14-magic-vision-dead"),
+                    TimeSpan.FromMinutes(10),
+                    ent);
+                break;
+            }
+        }
     }
 }
